+++ packages/ibm-products/src/components/Datagrid/Datagrid.test.js
@@ -167,6 +167,7 @@ const BasicUsage = ({ ...rest } = {}) => {
   const datagridState = useDatagrid({
     columns,
     data,
+    resizerAriaLabel: 'Custom column resizer label',
   });
 
   return <Datagrid datagridState={{ ...datagridState }} {...rest} />;
@@ -314,11 +315,12 @@ const TenThousandEntries = ({ ...rest } = {}) => {
       columns,
       data,
       rowSize: 'lg',
+      ...rest,
     },
     useInfiniteScroll
   );
 
-  return <Datagrid datagridState={{ ...datagridState }} {...rest} />;
+  return <Datagrid datagridState={datagridState} />;
 };
 
 const IsHoverOnRow = () => {
@@ -633,17 +635,31 @@ const SelectableRow = ({ ...rest } = {}) => {
 };
 
 const SortableColumns = ({ ...rest } = {}) => {
-  const columns = React.useMemo(() => defaultHeader, []);
+  const columns = React.useMemo(
+    () => [
+      ...defaultHeader,
+      {
+        Header: 'Someone 11',
+        accessor: 'someone11',
+        disableSortBy: true,
+      },
+    ],
+    []
+  );
   const [data] = useState(makeData(10));
   const datagridState = useDatagrid(
     {
       columns,
       data,
+      ascendingSortableLabelText: 'ascending',
+      descendingSortableLabelText: 'descending',
+      defaultSortableLabelText: 'none',
+      ...rest,
     },
     useSortableColumns
   );
 
-  return <Datagrid datagridState={{ ...datagridState }} {...rest} />;
+  return <Datagrid datagridState={datagridState} />;
 };
 
 const newPersonWithTwoLines = () => {
@@ -695,19 +711,20 @@ const TopAlignment = forwardRef(({ ...rest }, ref) => {
   return <Datagrid ref={ref} datagridState={{ ...datagridState }} {...rest} />;
 });
 
-const ClickableRow = ({ ...rest } = {}) => {
+const ClickableRow = ({ onRowClickFn, ...rest } = {}) => {
   const columns = React.useMemo(() => defaultHeader, []);
   const [data] = useState(makeData(10));
   const datagridState = useDatagrid(
     {
       columns,
       data,
-      onRowClick: (row) => alert(`Clicked ${row.id}`),
+      onRowClick: onRowClickFn,
+      ...rest,
     },
     useOnRowClick
   );
 
-  return <Datagrid datagridState={{ ...datagridState }} {...rest} />;
+  return <Datagrid datagridState={datagridState} />;
 };
 
 const InfiniteScroll = () => {
@@ -941,6 +958,10 @@ describe(componentName, () => {
     });
     const { keyboard, tab } = user;
     render(<BasicUsage data-testid={dataTestId} />);
+    const resizerInput = screen.getAllByLabelText(
+      'Custom column resizer label'
+    );
+    expect(resizerInput.length).toEqual(defaultHeader.length);
     await click(screen.getByTestId(dataTestId));
     await tab();
     // Input range resizer now has focus
@@ -1181,7 +1202,9 @@ describe(componentName, () => {
   });
 
   it('renders Ten Thousand table entries', async () => {
-    render(<TenThousandEntries data-testid={dataTestId} />);
+    const { rerender } = render(
+      <TenThousandEntries data-testid={dataTestId} />
+    );
 
     const tableBody =
       screen.getAllByRole('rowgroup')[1].firstElementChild.firstElementChild;
@@ -1191,6 +1214,21 @@ describe(componentName, () => {
     expect(
       parseInt(tableBodyHeight) / 48 // 48 is default row height
     ).toEqual(10000);
+
+    rerender(
+      <TenThousandEntries
+        virtualHeight={400}
+        data-testid={dataTestId}
+        loadMoreThreshold={300}
+      />
+    );
+    const rowGroups = screen.getAllByRole('rowgroup');
+    const bodyRowGroup = rowGroups[1];
+    const virtualScrollingElement = bodyRowGroup.firstElementChild;
+    fireEvent.scroll(virtualScrollingElement, { target: { scrollY: 5000 } });
+    expect(virtualScrollingElement.scrollLeft).toEqual(
+      bodyRowGroup.previousElementSibling.scrollLeft
+    );
   });
 
   it('With Pagination', async () => {
@@ -1205,26 +1243,46 @@ describe(componentName, () => {
   });
 
   it('Clickable Row', async () => {
-    const alertMock = jest.spyOn(window, 'alert');
-    render(<ClickableRow data-testid={dataTestId} />);
-    const node = screen
-      .getByRole('table')
-      .getElementsByTagName('tbody')[0]
-      .getElementsByTagName('tr')[0];
+    const onRowClickFn = jest.fn();
+    const { rerender } = render(
+      <ClickableRow onRowClickFn={onRowClickFn} data-testid={dataTestId} />
+    );
+    const rows = screen.getAllByRole('row');
+    const bodyRows = rows.filter(
+      (r) =>
+        !r.classList.contains('c4p--datagrid__head') &&
+        !r.classList.contains('c4p--datagrid__expanded-row')
+    );
 
-    fireEvent.click(node);
+    const firstBodyRow = bodyRows[0];
 
-    setTimeout(() => {
-      expect(alertMock).toHaveBeenCalledTimes(2);
-    }, 1000);
+    fireEvent.click(firstBodyRow);
+    expect(onRowClickFn).toHaveBeenCalledTimes(1);
 
-    fireEvent.keyDown(node, { key: 'Enter' });
-
-    setTimeout(() => {
-      expect(alertMock).toHaveBeenCalledTimes(2);
-    }, 1000);
-
-    fireEvent.keyDown(node, { key: '1' });
+    rerender(
+      <ClickableRow
+        isFetching
+        onRowClickFn={onRowClickFn}
+        data-testid={dataTestId}
+      />
+    );
+    const newRows = screen.getAllByRole('row');
+    const newBodyRows = newRows.filter(
+      (r) =>
+        !r.classList.contains('c4p--datagrid__head') &&
+        !r.classList.contains('c4p--datagrid__expanded-row')
+    );
+    const newBodyRow = newBodyRows[0];
+    newBodyRow.focus();
+    const user = userEvent.setup({
+      advanceTimers: jest.advanceTimersByTime,
+    });
+    const { keyboard } = user;
+    await keyboard('{Enter}');
+    expect(onRowClickFn).toHaveBeenCalledTimes(2);
+    newBodyRow.focus();
+    await keyboard('{Shift}');
+    expect(onRowClickFn).toHaveBeenCalledTimes(2);
   });
 
   function completeHoverOperation(rowNumber) {
@@ -1432,7 +1490,7 @@ describe(componentName, () => {
     fireEvent.click(rowExpanderCollapse);
   }
 
-  it('Expanded Row', () => {
+  it('should render with expandable rows and test by toggling the row open and closed', () => {
     render(<ExpandedRow data-testid={dataTestId} />);
     clickRow(1);
     clickRow(4);
@@ -1775,9 +1833,9 @@ describe(componentName, () => {
   const centerAlignedColumnsData = [
     ...defaultHeader.slice(0, 3),
     {
-      Header: 'Age',
+      Header: () => <span>Age</span>,
       accessor: 'age',
-      centerAlignedColumn: true,
+      rightAlignedColumn: true,
       disableSortBy: true,
     },
     {
@@ -1795,7 +1853,8 @@ describe(componentName, () => {
         data,
       },
       useColumnRightAlign,
-      useColumnCenterAlign
+      useColumnCenterAlign,
+      useSortableColumns
     );
 
     return <Datagrid datagridState={datagridState} />;
@@ -1850,9 +1909,7 @@ describe(componentName, () => {
     );
     const bodyAgeCell = bodyRows[0].childNodes[ageColIndex].firstChild;
     const bodyVisitsCell = bodyRows[0].childNodes[visitsColIndex].firstChild;
-    expect(bodyAgeCell).toHaveClass(
-      `${blockClass}__center-align-cell-renderer`
-    );
+    expect(bodyAgeCell).toHaveClass(`${blockClass}__right-align-cell-renderer`);
     expect(bodyAgeCell).toHaveClass(`sortDisabled`);
     expect(bodyVisitsCell).toHaveClass(
       `${blockClass}__center-align-cell-renderer`
@@ -1959,26 +2016,34 @@ describe(componentName, () => {
     );
   });
 
-  it('Sortable Columns', async () => {
-    render(<SortableColumns data-testid={dataTestId}></SortableColumns>);
+  it('should render sortable columns and toggle between sortable states for all column headers', () => {
+    const { click } = userEvent;
+    render(<SortableColumns data-testid={dataTestId} />);
 
-    const headerRow = screen
-      .getByRole('table')
-      .getElementsByTagName('thead')[0]
-      .getElementsByTagName('tr')[0];
+    const rows = screen.getAllByRole('row');
+    const headerRow = rows[0];
+    const columnHeaders = within(headerRow).getAllByRole('columnheader');
 
-    for (var i = 0; i < headerRow.getElementsByTagName('th').length - 1; i++) {
-      fireEvent.click(
-        headerRow
-          .getElementsByTagName('th')
-          .item(i)
-          .getElementsByTagName('div')[0]
-          .getElementsByTagName('button')[0]
+    Array.from(columnHeaders).map(async (colHeader, index) => {
+      // The last column definition opts out of sorting by specifying `disableSortBy`
+      // so we should not include testing for the last column header
+      if (index === defaultHeader.length) {
+        return;
+      }
+      const sortableColumnHeaderButton = within(colHeader).getByRole('button');
+      await click(sortableColumnHeaderButton);
+      expect(sortableColumnHeaderButton.getAttribute('aria-sort')).toEqual(
+        'ascending'
       );
-      expect(headerRow.getElementsByTagName('th')[i].classList[2]).toEqual(
-        'c4p--datagrid__isSorted'
+      await click(sortableColumnHeaderButton);
+      expect(sortableColumnHeaderButton.getAttribute('aria-sort')).toEqual(
+        'descending'
       );
-    }
+      await click(sortableColumnHeaderButton);
+      expect(sortableColumnHeaderButton.getAttribute('aria-sort')).toEqual(
+        'none'
+      );
+    });
   });
 
   it('Customizing Columns', async () => {
@@ -2007,100 +2072,9 @@ describe(componentName, () => {
     const ref = React.createRef();
     render(<TopAlignment ref={ref} data-testid={dataTestId} />);
 
-    const alertMock = jest.spyOn(window, 'alert');
-    expect(screen.getByRole('table').classList[2]).toEqual(
-      'c4p--datagrid__vertical-align-top'
+    expect(screen.getByRole('table')).toHaveClass(
+      `${blockClass}__vertical-align-top`
     );
-
-    const topAlignmentRows = screen
-      .getByRole('table')
-      .getElementsByTagName('tbody')[0]
-      .getElementsByTagName('tr');
-
-    const allRowsCheckBox = screen
-      .getByRole('table')
-      .getElementsByTagName('thead')[0]
-      .getElementsByTagName('tr')[0]
-      .getElementsByTagName('div')[0]
-      .getElementsByTagName('th')[0]
-      .getElementsByTagName('div')[0]
-      .getElementsByTagName('input')[0];
-
-    fireEvent.click(allRowsCheckBox);
-
-    for (var i = 0; i < topAlignmentRows.length; i++) {
-      expect(topAlignmentRows[i].classList[1]).toEqual(
-        `${carbon.prefix}--data-table--selected`
-      );
-    }
-
-    fireEvent.click(allRowsCheckBox);
-
-    for (var j = 0; j < topAlignmentRows.length; j++) {
-      fireEvent.click(
-        topAlignmentRows[j]
-          .getElementsByTagName('td')[0]
-          .getElementsByTagName('div')[0]
-          .getElementsByTagName('input')[0]
-      );
-      expect(topAlignmentRows[j].classList[1]).toEqual(
-        `${carbon.prefix}--data-table--selected`
-      );
-    }
-
-    fireEvent.click(allRowsCheckBox);
-
-    expect(
-      document.getElementsByClassName(`${carbon.prefix}--search-input`)[0]
-    ).toBeDefined();
-
-    expect(screen.getByText(/left panel/i)).toHaveClass(
-      `${carbon.prefix}--tooltip-content`
-    );
-
-    const rowHeightButton = screen.getByRole('button', {
-      name: /Row settings/i,
-    });
-    fireEvent.click(rowHeightButton);
-
-    const rowSizeDropDown = [
-      'Extra large',
-      'Large (default)',
-      'Medium',
-      'Extra small',
-    ];
-    const rowSize = ref.current.querySelector(
-      `.c4p--datagrid__row-size-dropdown`
-    );
-
-    for (var k = 0; k < rowSize; k++) {
-      expect(
-        document
-          .getElementsByClassName('c4p--datagrid__row-size-dropdown')[0]
-          .getElementsByTagName('div')[0]
-          .getElementsByTagName('fieldset')[0]
-          .getElementsByTagName('div')
-          .item(k)
-          .getElementsByTagName('label')[0]
-          .getElementsByTagName('span')[1].textContent
-      ).toEqual(rowSizeDropDown[k]);
-    }
-
-    fireEvent.click(
-      document
-        .getElementsByClassName('c4p--datagrid__table-toolbar')[0]
-        .getElementsByTagName('section')[0]
-        .getElementsByTagName('button')[0]
-    );
-    expect(alertMock).toHaveBeenCalledTimes(1);
-
-    const refreshButton = screen.getByLabelText('Refresh');
-    fireEvent.click(refreshButton);
-    expect(alertMock).toHaveBeenCalledTimes(2);
-
-    const downloadButton = screen.getByLabelText('Download CSV');
-    fireEvent.click(downloadButton);
-    expect(alertMock).toHaveBeenCalledTimes(3);
   });
 
   const getOverflowMenuItems = () => {