+++ packages/ibm-products/src/components/Datagrid/Datagrid/DatagridHeaderRow.js
@@ -31,6 +31,7 @@ const getAccessibilityProps = (header) => {
 };
 
 const HeaderRow = (datagridState, headRef, headerGroup) => {
+  const { resizerAriaLabel } = datagridState;
   // Used to measure the height of the table and uses that value
   // to display a vertical line to indicate the column you are resizing
   useEffect(() => {
@@ -68,6 +69,17 @@ const HeaderRow = (datagridState, headRef, headerGroup) => {
   }, [datagridState.rowSize, datagridState.tableId, datagridState]);
 
   const [incrementAmount] = useState(2);
+
+  const handleOnMouseDownResize = (event, resizeProps) => {
+    const { onMouseDown } = { ...resizeProps() };
+    // When event.button is 2, that is a right click
+    // and we do not want to resize
+    if (event.button === 2 || event.ctrlKey) {
+      event.target.blur();
+      return;
+    }
+    onMouseDown?.(event);
+  };
 
   return (
     <TableRow
@@ -114,6 +126,9 @@ const HeaderRow = (datagridState, headRef, headerGroup) => {
                 <>
                   <input
                     {...header.getResizerProps()}
+                    onMouseDown={(event) =>
+                      handleOnMouseDownResize(event, header.getResizerProps)
+                    }
                     onKeyDown={(event) => {
                       const { key } = event;
                       if (key === 'ArrowLeft' || key === 'ArrowRight') {
@@ -161,7 +176,7 @@ const HeaderRow = (datagridState, headRef, headerGroup) => {
                     className={cx(`${blockClass}__col-resizer-range`)}
                     type="range"
                     defaultValue={originalCol.width}
-                    aria-label="Resize column"
+                    aria-label={resizerAriaLabel || 'Resize column'}
                   />
                   <span className={`${blockClass}__col-resize-indicator`} />
                 </>