+++ packages/ibm-products/src/global/js/package-settings.js
@@ -73,6 +73,7 @@ const defaults = {
 
     // Novice to pro components not yet reviewed and released:
     Checklist: false,
+    InlineTip: false,
   },
 
   // feature level flags